Menambahkan 3 tabel kebijakan + user

+++ app/Http/Controllers/PeriodeController.php
@@ -9,10 +9,16 @@ use App\Models\User;
 use App\Http\Requests\StorePeriodeRequest;
 use App\Http\Requests\UpdatePeriodeRequest;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 
 class PeriodeController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -20,6 +26,8 @@ class PeriodeController extends Controller
      */
     public function index(Request $request)
     {
+        //$this->authorize('is_admin');
+        if(Gate::denies('is_OPD')) {
         
         $pagination = 10;
         // $periode = periodeModel::all()->OrderBy('tahun', 'asc');
@@ -27,6 +35,7 @@ class PeriodeController extends Controller
         $periode = PeriodeModel::OrderBy('tahun', 'asc')->get();
         return view('periode.index', ['periode'=>$periode])->with('i', ($request->input('page',1)-1)* $pagination);
     }
+}
 
     /**
      * Show the form for creating a new resource.
@@ -35,6 +44,7 @@ class PeriodeController extends Controller
      */
     public function create()
     {
+        $this->authorize('is_admin');
         $periode = PeriodeModel::all();
         return view('periode.create',['users'=> User::all()], compact('periode'));
 
@@ -48,6 +58,7 @@ class PeriodeController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('is_admin');
         $arrayValidation = [
             'tahun' => 'required|integer',
             'mulai' => 'required|date_format:Y-m-d\TH:i',
@@ -94,6 +105,7 @@ class PeriodeController extends Controller
      */
     public function edit(PeriodeModel $periode)
     {
+        $this->authorize('is_admin');
         $users = User::get();
         return view('periode.edit',compact(['users', 'periode']));
 
@@ -110,7 +122,7 @@ class PeriodeController extends Controller
     {  
 
     // $data=$request->all();
-
+    $this->authorize('is_admin');
     $arrayValidation = [
         'tahun' => 'required|integer',
         'mulai' => 'required|date_format:Y-m-d\TH:i',
@@ -144,12 +156,14 @@ class PeriodeController extends Controller
      */
     public function destroy(PeriodeModel $periode)
     {
+        $this->authorize('is_admin');
         $periode->delete($periode->id);
         return redirect()->route('periode.index')->with('success', 'data berhasil dihapus!');
     }
 
     public function changeStatus(Request $request)
         {
+            $this->authorize('is_admin');
             $periode = PeriodeModel::find($request->id);
             $periode->status = $request->status;
             $periode->save();
